update for pfazure
